+++ pkg/layered/delta.go
@@ -1,4 +1,4 @@
-package layers
+package layered
 
 import (
 	"fmt"
